1.New Architecture 2.add View 3.add Widgets
