+++ lib/features/crypto_list/view/crypto_list_screen.dart
@@ -1,0 +1,34 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_svg/svg.dart';
+import '../../../main.dart';
+
+
+class CryptoListScreen extends StatefulWidget {
+  const CryptoListScreen({
+    super.key,
+  });
+
+  @override
+  State<CryptoListScreen> createState() => _CryptoListScreenState();
+}
+
+class _CryptoListScreenState extends State<CryptoListScreen> {
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('CryptoCurrenciesList'),
+      ),
+      body: ListView.separated(
+        itemCount: 10,
+        separatorBuilder: (context, index) => const Divider(),
+        itemBuilder: (context, i) {
+    const coinName = 'Bitcoin';
+    return const CryptoCoinScreen(coinName : coinName);
+    },
+    ),
+    );
+}
+}
