+++ lib/features/crypto_list/view/view.dart
@@ -1,0 +1,1 @@
+export 'crypto_list_screen.dart';
