+++ lib/features/crypto_list/widgets/crypto_coin_title.dart
@@ -1,0 +1,37 @@
+
+
+import 'package:flutter/material.dart';
+import 'package:flutter_svg/svg.dart';
+
+class CryptoCoinTile extends StatelessWidget {
+  const CryptoCoinTile({
+    Key? key,
+    required this.coinName,
+  }) : super(key: key);
+  final String coinName;
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+    return ListTile(
+      leading: SvgPicture.asset('assets/svg/bitcoin_logo.svg',
+        height: 30,
+        width: 30,
+      ),
+      title: Text(
+        coinName,
+        style: theme.textTheme.bodyMedium,
+      ),
+      subtitle: Text(
+        '20000\$',
+        style: theme.textTheme.labelSmall,
+      ),
+      trailing: const Icon(
+          Icons.arrow_forward_ios),
+      onTap: () {
+        Navigator.of(context).pushNamed(
+          '/coin', arguments: coinName,);
+      },
+    );
+  }
+}
