+++ lib/features/crypto_list/widgets/crypto_list.dart
@@ -1,0 +1,1 @@
+export '../view/view.dart';
