+++ lib/features/crypto_list/widgets/widgets.dart
@@ -1,0 +1,1 @@
+export 'crypto_coin_title.dart';
